requirements.txt
requests
beautifulsoup4
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib